handle_input added to main.py and bug fixing

+++ ir_emitter.py
@@ -3,6 +3,7 @@ import esp32
 import time
 import sys
 import ujson
+import gc
 
 ir_led = PWM(Pin(33), freq=38000, duty=0)
 
@@ -10,7 +11,7 @@ def send_signal(data):
     print("Sende neuen NEC-Code...")
     for level, duration in data:
         if level == 0:
-            # Bei NEC bedeutet der 0-Level vom Receiver meistens: LED AN
+            #0 meistens = an
             ir_led.duty(512)
         else:
             ir_led.duty(0)
@@ -22,17 +23,20 @@ def send_signal(data):
 
 
 
-def get_ir():
-    # auf serielle Daten warten
-    line = sys.stdin.readline()
+def get_ir(line):
+    #auf serielle Daten warten
     print("read line"+line)
     if not line:
         return
-    try:
-        # Pulsfolge als JSON empfangen
-        pulses = ujson.loads(line)
-        for i in range (3):
+    if line.startswith('['):
+        try:
+            #Pulsfolge als JSON empfangen
+            gc.collect()
+            pulses = ujson.loads(line)
             send_signal(pulses)
-        print("IR Signal gesendet")
-    except Exception as e:
-        print("Fehler beim Senden:", e)
+            print("IR Signal gesendet")
+            gc.collect()
+        except Exception as e:
+            print("Fehler beim Senden:", e)
+        else:
+            pass
